ai-image-app/backend/requirements_video.txt
# Ultra-Realistic Video Generation System Requirements
# Core AI/ML Libraries
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.24.0
transformers>=4.30.0
accelerate>=0.20.0

# Video Processing
opencv-python>=4.8.0
imageio>=2.25.0
imageio-ffmpeg>=0.4.8
av>=10.0.0

# Image Processing
Pillow>=9.0.0
numpy>=1.24.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
requests>=2.28.0
tqdm>=4.64.0
safetensors>=0.3.0
huggingface-hub>=0.16.0

# Optional: For better performance
xformers>=0.0.20  # Memory efficient attention
bitsandbytes>=0.41.0  # Quantization support

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0 